backend/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP requests and API clients
httpx>=0.24.0
requests>=2.28.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
networkx>=3.0

# GTFS-RT parsing
gtfs-realtime-bindings>=1.0.0
protobuf>=4.21.0

# Database
sqlalchemy>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
